🧹 refactor: split complex unit test module into domain submodules
